refactor(shared/chat_protocol): _ChatEventBuffer 를 별 모듈로 분리 (SRP)
session_runtime.py 안에 정의돼 있던 _ChatEventBuffer 를 _chat_event_buffer.py
로 분리. SessionRuntime / SessionRegistry / TTL sweeper 와 deque 기반 버퍼
자료구조의 책임이 다름.

- _chat_event_buffer.py: deque + anyio.Event 기반 message-aware 버퍼
- session_runtime.py: SessionRuntime + SessionRegistry + sweeper

underscore prefix 유지 — implementation detail 로 외부 미노출.
